chore: component name modified & error state deleted
