+++ src/components/assetLibrary/MediaAssetThumbnail.vue
